backend/requirements.txt
asgiref
django
django-redis
redis
sqlparse
dotenv
djangorestframework
django-cors-headers
psycopg2-binary
channels
channels-redis
requests
geopy
python-dotenv
django-debug-toolbar
pytest
pytest-django
gunicorn
passiogo_up

